refactor: fixing and identifying some potential bugs - add development environment connection string - reducing some dependencies - renaming and restructuring
