+++ src/Api/Controllers/ApiControllerBase.cs
@@ -4,8 +4,8 @@ using Microsoft.AspNetCore.Mvc;
 namespace Api.Controllers;
 
 [ApiController]
-[Route("[controller]")]
-public class ApiControllerBase : ControllerBase
+[Route("api/v1/[controller]")]
+public abstract class ApiControllerBase : ControllerBase
 {
     private ISender? _mediator;
     protected ISender Mediator => _mediator ??= HttpContext.RequestServices.GetRequiredService<ISender>();
