+++ src/Api/Controllers/UsersController.cs
@@ -1,15 +1,14 @@
 using Application.Common.Models;
-using Application.Users.Commands.CreateUserCommand;
+using Application.Users.Commands.CreateUser;
 using Application.Users.Queries;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Api.Controllers;
 
-[Route("api/v1/[controller]")]
 public class UsersController : ApiControllerBase
 {
     [HttpPost]
-    public async Task<ActionResult<Result<UserDto>>> CreateUser(CreateUserCommand command)
+    public async Task<ActionResult<Result<UserDto>>> CreateUser([FromBody] CreateUserCommand command)
     {
         var result = await Mediator.Send(command);
         return Ok(Result<UserDto>.Succeed(result));
