+++ src/Api/appsettings.Development.json
@@ -1,4 +1,8 @@
 {
+  "DatabaseSettings": {
+    "DbProvider": "PostgreSQL",
+    "ConnectionString": "Server=localhost;Port=5432;Database=mydb;User ID=profiletest;Password=supasecured;"
+  },
   "Serilog" : {
     "MinimumLevel" : {
       "Default": "Debug",
